+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -10,7 +10,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import kitchenpos.domain.BaseEntity;
-import kitchenpos.domain.Price;
+import kitchenpos.product.domain.Price;
 
 @Entity
 public class Menu extends BaseEntity {
