+++ src/test/java/kitchenpos/domain/OrderTableTest.java
@@ -8,6 +8,7 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
+import kitchenpos.product.domain.Price;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
